requirement.txt
requests==2.31.0
anthropic==0.3.0
geopy==2.3.0
python-dotenv==1.0.0
